requirements.txt
appdirs==1.4.4
apturl==0.5.2
asgiref==3.4.1
attrs==19.3.0
autopep8==1.5.7
blinker==1.4
Brlapi==0.7.0
cached-property==1.5.1
certifi==2019.11.28
chardet==3.0.4
chrome-gnome-shell==0.0.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.2.6
django-crispy-forms==1.11.2
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
drf-yasg==1.20.0
entrypoints==0.3
filelock==3.0.12
httplib2==0.14.0
idna==2.8
importlib-metadata==1.5.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.12.0
macaroonbakery==1.3.1
Markdown==3.3.4
MarkupSafe==2.0.1
more-itertools==4.2.0
netifaces==0.10.4
notifications==0.3.2
oauthlib==3.1.0
olefile==0.46
packaging==21.0
pbr==5.5.1
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
psycopg2==2.8.6
PyAudio==0.2.11
pycairo==1.16.2
pycodestyle==2.7.0
pycups==1.9.73
PyGObject==3.36.0
PyJWT==2.2.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.7
pyping==0.0.6
pyRFC3339==1.1
pyrsistent==0.15.5
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytube==10.8.2
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
ruamel.yaml==0.17.13
ruamel.yaml.clib==0.2.6
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
SpeechRecognition==3.8.1
sqlparse==0.2.4
stevedore==3.3.0
systemd-python==234
texttable==1.6.2
toml==0.10.2
ubuntu-advantage-tools==27.0
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.25.8
vboxapi==1.0
virtualenv==20.4.3
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
wadllib==1.3.3
websocket-client==0.53.0
xkit==0.0.0
zipp==1.0.0
